#80 Replace Bootstrap v2.3.2 with v3.0.0-rc1
This seriously damages Backstage, this damage will be fixed during this
ticket and the Backstage redesign. Also: using unstable releases of
Bootstrap can be excepted during the beta stage of ModernBB, however, IF
ModernBB 2.5.0 should be finished before Bootstrap 3.0.0 is, we will
release it with an unstable Bootstrap release and fix that issue so fast
as we can in v2.5.1.