#88 Stop supporting updating for old versions
We have to keep on track and for ModernBB 2.0, that means that we've to
dump support for updating from PunBB 1.2, 1.3 and FluxBB 1.2, 1.3. It
seems to be a very good point, since the update file is now 42% lighter.
This commit also updates the design of the update message and fixes the
design for the "Finished" message.